CDEC-431: Move some update code up into the lib package
