+++ lib/src/Pos/Listener/Update.hs
@@ -5,19 +5,55 @@
 -- It's not about the network at all.
 
 module Pos.Listener.Update
-       ( handleProposal
+       ( UpdateMode
+       , handleProposal
        , handleVote
        ) where
 
 import           Universum
 
 import           Formatting (build, sformat, (%))
+import           Mockable (MonadMockable)
 import           System.Wlog (WithLogger, logNotice, logWarning)
+import           UnliftIO (MonadUnliftIO)
 
 import           Pos.Core (ProtocolMagic)
 import           Pos.Core.Update (UpdateProposal (..), UpdateVote (..))
+import           Pos.DB.Class (MonadDB, MonadGState)
+import           Pos.Infra.Recovery.Info (MonadRecoveryInfo)
+import           Pos.Infra.Reporting (MonadReporting)
+import           Pos.Infra.Shutdown.Class (HasShutdownContext)
+import           Pos.Infra.Slotting (MonadSlots)
+import           Pos.Infra.StateLock (StateLock)
+import           Pos.Lrc.Context (HasLrcContext)
+import           Pos.Update.Configuration (HasUpdateConfiguration)
+import           Pos.Update.Context (UpdateContext)
 import           Pos.Update.Logic.Local (processProposal, processVote)
-import           Pos.Update.Mode (UpdateMode)
+import           Pos.Update.Params (UpdateParams)
+import           Pos.Util.Util (HasLens (..))
+
+
+
+
+type UpdateMode ctx m
+    = ( WithLogger m
+      , MonadMockable m
+      , MonadIO m
+      , MonadUnliftIO m
+      , MonadMask m
+      , MonadGState m
+      , MonadDB m
+      , MonadReader ctx m
+      , HasLrcContext ctx
+      , HasLens UpdateContext ctx UpdateContext
+      , HasLens UpdateParams ctx UpdateParams
+      , HasLens StateLock ctx StateLock
+      , HasShutdownContext ctx
+      , HasUpdateConfiguration
+      , MonadReporting m
+      , MonadRecoveryInfo m
+      , MonadSlots ctx m
+      )
 
 handleProposal
     :: forall ctx m . UpdateMode ctx m
