+++ lib/src/Pos/Logic/Full.hs
@@ -34,11 +34,15 @@ import qualified Pos.DB.BlockIndex as DB (getHeader, getTipHeader)
 import           Pos.DB.Class (MonadBlockDBRead, MonadDBRead, MonadGState (..),
                      SerializedBlock)
 import qualified Pos.DB.Class as DB (MonadDBRead (dbGetSerBlock))
-import           Pos.Listener.Delegation (DlgListenerConstraint)
-import qualified Pos.Listener.Delegation as Delegation (handlePsk)
 import qualified Pos.GState.BlockExtra as DB (resolveForwardLink, streamBlocks)
 import           Pos.Infra.Slotting (MonadSlots)
 import           Pos.Infra.Util.JsonLog.Events (JLEvent)
+import           Pos.Listener.Delegation (DlgListenerConstraint)
+import qualified Pos.Listener.Delegation as Delegation (handlePsk)
+import           Pos.Listener.Txp (TxpMode)
+import qualified Pos.Listener.Txp as Txp (handleTxDo)
+import           Pos.Listener.Update (UpdateMode)
+import qualified Pos.Listener.Update as Update (handleProposal, handleVote)
 import           Pos.Logic.Types (KeyVal (..), Logic (..))
 import           Pos.Recovery (MonadRecoveryInfo)
 import qualified Pos.Recovery as Recovery
@@ -54,13 +58,8 @@ import           Pos.Ssc.Toss (SscTag (..), TossModifier, tmCertificates,
 import           Pos.Ssc.Types (ldModifier)
 import           Pos.Txp (MemPool (..))
 import           Pos.Txp.MemState (getMemPool, withTxpLocalData)
-import           Pos.Listener.Txp (TxpMode)
-import qualified Pos.Listener.Txp as Txp (handleTxDo)
 import qualified Pos.Update.Logic.Local as Update (getLocalProposalNVotes,
                      getLocalVote, isProposalNeeded, isVoteNeeded)
-import           Pos.Update.Mode (UpdateMode)
-import qualified Pos.Listener.Update as Update (handleProposal,
-                     handleVote)
 import           Pos.Util.Util (HasLens (..))
 
 -- The full logic layer uses existing pieces from the former monolithic
