+++ lib/src/Pos/Network/Update/Download.hs
@@ -32,10 +32,10 @@ import           Pos.Core.Update (SoftwareVersion (..), UpdateData (..),
 import           Pos.Crypto (Hash, castHash, hash)
 import           Pos.Exception (reportFatalError)
 import           Pos.Infra.Reporting (reportOrLogW)
+import           Pos.Listener.Update (UpdateMode)
 import           Pos.Update.Configuration (curSoftwareVersion, ourSystemTag)
 import           Pos.Update.Context (UpdateContext (..))
 import           Pos.Update.DB.Misc (isUpdateInstalled)
-import           Pos.Update.Mode (UpdateMode)
 import           Pos.Update.Params (UpdateParams (..))
 import           Pos.Update.Poll.Types (ConfirmedProposalState (..))
 import           Pos.Util.Concurrent (withMVar)
