+++ lib/src/Pos/Worker/Update.hs
@@ -19,12 +19,12 @@ import           Pos.Infra.Recovery.Info (recoveryCommGuard)
 import           Pos.Infra.Shutdown (triggerShutdown)
 import           Pos.Infra.Slotting.Util (ActionTerminationPolicy (..),
                      OnNewSlotParams (..), defaultOnNewSlotParams, onNewSlot)
+import           Pos.Listener.Update (UpdateMode)
+import           Pos.Network.Update.Download (downloadUpdate)
 import           Pos.Update.Configuration (curSoftwareVersion)
 import           Pos.Update.Context (UpdateContext (..))
 import           Pos.Update.DB (getConfirmedProposals)
-import           Pos.Network.Update.Download (downloadUpdate)
 import           Pos.Update.Logic.Local (processNewSlot)
-import           Pos.Update.Mode (UpdateMode)
 import           Pos.Update.Poll.Types (ConfirmedProposalState (..))
 import           Pos.Util.Util (lensOf)
