+++ tools/src/launcher/Main.hs
@@ -76,8 +76,8 @@ import           Pos.Infra.Reporting.Http (sendReport)
 import           Pos.Infra.Reporting.Wlog (compressLogs, retrieveLogFiles)
 import           Pos.Launcher (HasConfigurations, withConfigurations)
 import           Pos.Launcher.Configuration (ConfigurationOptions (..))
+import           Pos.Network.Update.Download (installerHash)
 import           Pos.ReportServer.Report (ReportType (..))
-import           Pos.Update (installerHash)
 import           Pos.Update.DB.Misc (affirmUpdateInstalled)
 import           Pos.Util (HasLens (..), directory, logException,
                      postfixLFields)
